backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.31
apscheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
resend==2.4.0
gunicorn==22.0.0
